requirements.txt
aiogram~=3.16.0
pytesseract~=0.3.13
pillow~=11.0.0
python-dotenv~=1.0.1